Renamed legacy namespace
