+++ Library/Compression/Integers/Gen2/IByteWriter.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace InvertedTomato.Compression.Integers.Legacy {
+namespace InvertedTomato.Compression.Integers.Gen2 {
 	public interface IByteWriter {
 		void WriteByte(Byte value);
 	}
