+++ Library/Compression/Integers/Gen2/Wrappers/StreamWrapper.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 
-namespace InvertedTomato.Compression.Integers.Legacy {
+namespace InvertedTomato.Compression.Integers.Gen2 {
 	public class StreamWrapper : IByteReader, IByteWriter {
 		private readonly Stream Underlying;
